+++ libnavigation-copilot/src/main/java/com/mapbox/navigation/copilot/GoingToBackground.kt
@@ -1,0 +1,6 @@
+package com.mapbox.navigation.copilot
+
+import androidx.annotation.Keep
+
+@Keep
+internal object GoingToBackground : EventDTO